[#138] feat: 카드정보데이터 가공 함수 분리
